[In progress] - Combining all homo transformations

+++ catkin_ws_Indep.Study/src/mapping_simulator/src/noise.cpp
@@ -10,22 +10,22 @@ Noise::Noise()
     std = 0.0;
 }
 
-Noise::Noise(double mean, double std)
+Noise::Noise(float mean, float std)
 {
     this->mean = mean;
     this->std = std;
 }
 
-void Noise::change_mean_std(double mean, double std)
+void Noise::change_mean_std(float mean, float std)
 {
     this->mean = mean;
     this->std = std;
 }
 
-double Noise::gaussian()
+float Noise::gaussian()
 {
     seed = std::chrono::system_clock::now().time_since_epoch().count();
     default_random_engine generator(seed);
-    normal_distribution<double> distribution(mean, std);
+    normal_distribution<float> distribution(mean, std);
     return distribution(generator);
 }
